\name{CWe1-package}
\alias{CWe1-package}
\alias{CWe1}
\docType{package}
\title{
\packageTitle{CWe1}
}
\description{
\packageDescription{CWe1}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CWe1}
\packageIndices{CWe1}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{CWe1}

Maintainer: \packageMaintainer{CWe1}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
